Undo trait fns
