+++ crates/utils/src/tracing/span_ext.rs
@@ -59,51 +59,19 @@ impl ToValue for i64 {
 ///
 /// This is a sealed trait. It and cannot be implemented outside of this module.
 pub trait OpenTelemetrySpanExt: private::Sealed {
-    fn set_parent<T>(&self, request: &http::Request<T>);
     fn set_attribute(&self, key: impl Into<Key>, value: impl ToValue);
-    fn set_http_attributes<T>(&self, request: &http::Request<T>);
     fn set_error(&self, err: &dyn std::error::Error);
-    fn context(&self) -> opentelemetry::Context;
 }
 
 impl<S> OpenTelemetrySpanExt for S
 where
     S: tracing_opentelemetry::OpenTelemetrySpanExt,
 {
-    /// Sets the parent context by extracting HTTP metadata from the request.
-    fn set_parent<T>(&self, request: &http::Request<T>) {
-        // Pull the open-telemetry parent context using the HTTP extractor. We could make a more
-        // generic gRPC extractor by utilising the gRPC metadata. However that
-        //     (a) requires cloning headers,
-        //     (b) we would have to write this ourselves, and
-        //     (c) gRPC metadata is transferred using HTTP headers in any case.
-        let otel_ctx = opentelemetry::global::get_text_map_propagator(|propagator| {
-            propagator.extract(&MetadataExtractor(&tonic::metadata::MetadataMap::from_headers(
-                request.headers().clone(),
-            )))
-        });
-        tracing_opentelemetry::OpenTelemetrySpanExt::set_parent(self, otel_ctx);
-    }
-
-    /// Returns the context of `Span`.
-    fn context(&self) -> opentelemetry::Context {
-        tracing_opentelemetry::OpenTelemetrySpanExt::context(self)
-    }
-
     /// Sets an attribute on `Span`.
     ///
     /// Implementations for `ToValue` should be added to this crate (miden-node-utils).
     fn set_attribute(&self, key: impl Into<Key>, value: impl ToValue) {
         tracing_opentelemetry::OpenTelemetrySpanExt::set_attribute(self, key, value.to_value());
-    }
-
-    /// Sets standard attributes to the `Span` based on an associated HTTP request.
-    fn set_http_attributes<T>(&self, request: &http::Request<T>) {
-        let remote_addr = request
-            .extensions()
-            .get::<tonic::transport::server::TcpConnectInfo>()
-            .and_then(tonic::transport::server::TcpConnectInfo::remote_addr);
-        OpenTelemetrySpanExt::set_attribute(self, "remote_addr", remote_addr);
     }
 
     /// Sets a status on `Span` based on an error.
@@ -124,25 +92,4 @@ where
 mod private {
     pub trait Sealed {}
     impl<S> Sealed for S where S: tracing_opentelemetry::OpenTelemetrySpanExt {}
-}
-
-/// Facilitates Open Telemetry metadata extraction for Tonic `MetadataMap`.
-struct MetadataExtractor<'a>(&'a tonic::metadata::MetadataMap);
-impl opentelemetry::propagation::Extractor for MetadataExtractor<'_> {
-    /// Get a value for a key from the `MetadataMap`.  If the value can't be converted to &str,
-    /// returns None
-    fn get(&self, key: &str) -> Option<&str> {
-        self.0.get(key).and_then(|metadata| metadata.to_str().ok())
-    }
-
-    /// Collect all the keys from the `MetadataMap`.
-    fn keys(&self) -> Vec<&str> {
-        self.0
-            .keys()
-            .map(|key| match key {
-                tonic::metadata::KeyRef::Ascii(v) => v.as_str(),
-                tonic::metadata::KeyRef::Binary(v) => v.as_str(),
-            })
-            .collect::<Vec<_>>()
-    }
 }
